feat(mysql): support mysql2 module
