+++ lib/instrumentation/index.js
@@ -9,7 +9,7 @@ var protocol = require('./protocol')
 var debug = require('debug')('elastic-apm')
 var request = require('../request')
 
-var MODULES = ['http', 'https', 'generic-pool', 'mongodb-core', 'pg', 'mysql', 'express', 'hapi', 'redis', 'ioredis', 'bluebird', 'knex', 'koa-router', 'ws', 'graphql', 'express-graphql', 'elasticsearch', 'handlebars']
+var MODULES = ['http', 'https', 'generic-pool', 'mongodb-core', 'pg', 'mysql', 'mysql2', 'express', 'hapi', 'redis', 'ioredis', 'bluebird', 'knex', 'koa-router', 'ws', 'graphql', 'express-graphql', 'elasticsearch', 'handlebars']
 
 module.exports = Instrumentation
