+++ lib/instrumentation/modules/mysql2.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = require('./mysql-common')('mysql2', '^1.0.0')
